switching to junit 5 removed powermock using an interface in the Logger
