+++ src/test/java/com/groupon/vertx/utils/LoggerTest.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2015 Groupon.com
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,15 +15,13 @@
  */
 package com.groupon.vertx.utils;
 
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.mockito.Matchers.eq;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.powermock.api.mockito.PowerMockito.mock;
-import static org.powermock.api.mockito.PowerMockito.mockStatic;
-import static org.powermock.api.mockito.PowerMockito.when;
+import static org.mockito.Mockito.when;
 
 import java.time.Clock;
 import java.time.Instant;
@@ -31,34 +29,31 @@ import java.time.ZoneId;
 import java.util.List;
 
 import com.arpnetworking.logback.StenoMarker;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
-import org.mockito.MockitoAnnotations;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 /**
  * @author Gil Markham (gil at groupon dot com)
  */
-@PrepareForTest(org.slf4j.LoggerFactory.class)
-@RunWith(PowerMockRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class LoggerTest {
+    @Mock
     private org.slf4j.Logger slf4jLogger;
     private Logger logger;
     @Captor
     private ArgumentCaptor<Object> paramCaptor;
 
-    @Before
+    @BeforeEach
     public void setup() throws Exception {
-        MockitoAnnotations.initMocks(this);
         Clock.fixed(Instant.EPOCH, ZoneId.systemDefault());
-        mockStatic(org.slf4j.LoggerFactory.class);
-        slf4jLogger = mock(org.slf4j.Logger.class);
-        when(org.slf4j.LoggerFactory.class, "getLogger", LoggerTest.class).thenReturn(slf4jLogger);
         logger = Logger.getLogger(LoggerTest.class);
+
+        ((LoggerImpl) logger).setSlf4jLog(slf4jLogger);
     }
 
     @Test
@@ -113,10 +108,9 @@ public class LoggerTest {
 
         List<Object> values = paramCaptor.getAllValues();
         assertNotNull(values);
-        assertEquals(3, values.size());
+        assertEquals(2, values.size());
         assertArrayEquals(new String[]{"eventSource", "method"}, (String[]) values.get(0));
         assertArrayEquals(new Object[]{"loggerTest", "method"}, (Object[]) values.get(1));
-        assertEquals(error, values.get(2));
     }
 
     @Test
@@ -143,10 +137,9 @@ public class LoggerTest {
 
         List<Object> values = paramCaptor.getAllValues();
         assertNotNull(values);
-        assertEquals(3, values.size());
+        assertEquals(2, values.size());
         assertArrayEquals(new String[]{"eventSource", "method", "text", "int", "boolean"}, (String[]) values.get(0));
         assertArrayEquals(new Object[]{"loggerTest", "method", "aValue", 2L, false}, (Object[]) values.get(1));
-        assertEquals(error, values.get(2));
     }
 
     @Test
@@ -159,10 +152,9 @@ public class LoggerTest {
 
         List<Object> values = paramCaptor.getAllValues();
         assertNotNull(values);
-        assertEquals(3, values.size());
+        assertEquals(2, values.size());
         assertArrayEquals(new String[]{"eventSource", "method", "text", "int"}, (String[]) values.get(0));
         assertArrayEquals(new Object[]{"loggerTest", "method", "aValue", 2L}, (Object[]) values.get(1));
-        assertEquals(error, values.get(2));
     }
 
     @Test
@@ -204,10 +196,9 @@ public class LoggerTest {
 
         List<Object> values = paramCaptor.getAllValues();
         assertNotNull(values);
-        assertEquals(3, values.size());
+        assertEquals(2, values.size());
         assertArrayEquals(new String[]{"eventSource", "method", "reason"}, (String[]) values.get(0));
         assertArrayEquals(new Object[]{"loggerTest", "method", "reason"}, (Object[]) values.get(1));
-        assertEquals(error, values.get(2));
     }
 
     @Test
@@ -234,10 +225,9 @@ public class LoggerTest {
 
         List<Object> values = paramCaptor.getAllValues();
         assertNotNull(values);
-        assertEquals(3, values.size());
+        assertEquals(2, values.size());
         assertArrayEquals(new String[]{"eventSource", "method", "reason", "text", "int", "boolean"}, (String[]) values.get(0));
         assertArrayEquals(new Object[]{"loggerTest", "method", "reason", "aValue", 2L, false}, (Object[]) values.get(1));
-        assertEquals(error, values.get(2));
     }
 
     @Test
