+++ src/test/java/com/groupon/vertx/utils/deployment/MultiVerticleDeploymentTest.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2015 Groupon.com
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,11 +15,11 @@
  */
 package com.groupon.vertx.utils.deployment;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.eq;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.when;
 
@@ -33,14 +33,13 @@ import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.file.FileSystem;
 import io.vertx.core.json.JsonObject;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 
 import com.groupon.vertx.utils.config.ConfigLoader;
@@ -100,12 +99,9 @@ public class MultiVerticleDeploymentTest {
     }
 
     private Answer<Object> invokeHandlerWithResult(final Future<String> result) {
-        return new Answer<Object>() {
-            @Override
-            public Object answer(InvocationOnMock invocationOnMock) throws Throwable {
-                handlerCaptor.getValue().handle(result);
-                return null;
-            }
+        return invocationOnMock -> {
+            handlerCaptor.getValue().handle(result);
+            return null;
         };
     }
 
@@ -119,7 +115,7 @@ public class MultiVerticleDeploymentTest {
                 .when(deployment).abort(any(Throwable.class));
     }
 
-    @Before
+    @BeforeEach
     public void setup() {
         MockitoAnnotations.initMocks(this);
 
@@ -139,23 +135,18 @@ public class MultiVerticleDeploymentTest {
         latch = new CountDownLatch(1);
     }
 
-    @After
+    @AfterEach
     public void ensureFinish() throws Exception {
         latch.await(TEST_TIMEOUT, TimeUnit.MILLISECONDS);
 
-        if (latch.getCount() != 0) {
-            fail("Timed out; test did not finish in " + TEST_TIMEOUT + "ms");
-        }
+        assertNotEquals(0, latch.getCount());
     }
 
     @Test
     public void testSuccess() {
-        multiVerticleDeployment.deploy(config).setHandler(new Handler<AsyncResult<Void>>() {
-            @Override
-            public void handle(AsyncResult<Void> result) {
-                assertTrue("Deployment should succeed", result.succeeded());
-                latch.countDown();
-            }
+        multiVerticleDeployment.deploy(config).setHandler(result -> {
+            assertTrue(result.succeeded(), "Deployment should succeed");
+            latch.countDown();
         });
     }
 
@@ -163,12 +154,9 @@ public class MultiVerticleDeploymentTest {
     public void testBadConfig() {
         config.remove("verticles");
 
-        multiVerticleDeployment.deploy(config).setHandler(new Handler<AsyncResult<Void>>() {
-            @Override
-            public void handle(AsyncResult<Void> result) {
-                assertTrue("Deployment should fail", result.failed());
-                latch.countDown();
-            }
+        multiVerticleDeployment.deploy(config).setHandler(result -> {
+            assertTrue(result.failed(), "Deployment should fail");
+            latch.countDown();
         });
 
     }
@@ -177,12 +165,9 @@ public class MultiVerticleDeploymentTest {
     public void testDeployFailure() {
         stubDeploymentDeployWithResult(Future.<String>failedFuture(new Exception("failure")));
 
-        multiVerticleDeployment.deploy(config).setHandler(new Handler<AsyncResult<Void>>() {
-            @Override
-            public void handle(AsyncResult<Void> result) {
-                assertTrue("Deployment should fail", result.failed());
-                latch.countDown();
-            }
+        multiVerticleDeployment.deploy(config).setHandler(result -> {
+            assertTrue(result.failed(), "Deployment should fail");
+            latch.countDown();
         });
 
     }
@@ -191,12 +176,9 @@ public class MultiVerticleDeploymentTest {
     public void testBadVerticleConfigA() {
         config.getJsonObject("verticles").getJsonObject(VERTICLE_NAME_A).remove("instances");
 
-        multiVerticleDeployment.deploy(config).setHandler(new Handler<AsyncResult<Void>>() {
-            @Override
-            public void handle(AsyncResult<Void> result) {
-                assertTrue("Deployment should fail", result.failed());
-                latch.countDown();
-            }
+        multiVerticleDeployment.deploy(config).setHandler(result -> {
+            assertTrue(result.failed(), "Deployment should fail");
+            latch.countDown();
         });
     }
 
@@ -204,12 +186,9 @@ public class MultiVerticleDeploymentTest {
     public void testBadVerticleConfigB() {
         config.getJsonObject("verticles").getJsonObject(VERTICLE_NAME_A).remove("class");
 
-        multiVerticleDeployment.deploy(config).setHandler(new Handler<AsyncResult<Void>>() {
-            @Override
-            public void handle(AsyncResult<Void> result) {
-                assertTrue("Deployment should fail", result.failed());
-                latch.countDown();
-            }
+        multiVerticleDeployment.deploy(config).setHandler(result -> {
+            assertTrue(result.failed(), "Deployment should fail");
+            latch.countDown();
         });
     }
 
@@ -219,7 +198,7 @@ public class MultiVerticleDeploymentTest {
         try {
             multiVerticleDeployment.deploy(config);
         } catch (IllegalStateException e) {
-            assertNotNull("Throws IllegalStateException when trying to deploy a second time with the same object", e);
+            assertNotNull(e, "Throws IllegalStateException when trying to deploy a second time with the same object");
         }
         latch.countDown();
     }
